+++ examples/configuration/graphviz.php
@@ -1,34 +1,20 @@
 <?php declare(strict_types=1);
 
-require_once __DIR__ . '/classmap.php';
+require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
 
 use Bartlett\GraphUml\Generator\GraphVizGenerator;
 use Bartlett\UmlWriter\Generator\GeneratorFactory;
 use Bartlett\UmlWriter\Service\ClassDiagramRenderer;
 
-use Name\Space\MyGeneratorFactory;
-
 use Symfony\Component\Finder\Finder;
 
 // path to directory where to find PHP source code
-$dataSource = dirname(__DIR__, 2) . '/src';
+$dataSource1 = dirname(__DIR__, 2) . '/src/Config';
+$dataSource2 = dirname(__DIR__, 2) . '/src/Service';
 
 $finder = new Finder();
-$finder->in($dataSource)->name('*.php');
-
-$generatorFactory = new MyGeneratorFactory('mygenerator');
-// creates instance of Name\Space\MyGenerator
-$generator = $generatorFactory->getGenerator();
-
-$renderer = new ClassDiagramRenderer();
-// generates UML class diagram of all objects found in dataSource
-$script = $renderer($finder, $generator);
-// show UML diagram statements
-echo $script;
-
-
-$finder = new Finder();
-$finder->in(__DIR__)->name('*.php')->notName('graphviz.php');
+$finder->in($dataSource1)->name('*.php');
+$finder->append((new Finder())->in($dataSource2)->name('ConfigurationHandler.php'));
 
 $generatorFactory = new GeneratorFactory('graphviz');
 
@@ -44,14 +30,16 @@ $options = [
     'node.style' => 'filled',
 ];
 // generates UML class diagram of all objects found in dataSource (in graphviz format)
-$renderer($finder, $generator, $options);
+$script = $renderer($finder, $generator, $options);
+// show UML diagram statements
+echo $script;
 
 // default format is PNG, change it to SVG
 $generator->setFormat('svg');
 
 if (isset($argv[1])) {
     // target folder provided
-    $cmdFormat = '%E -T%F %t -o ' . rtrim($argv[1], DIRECTORY_SEPARATOR) . '/generator.graphviz.%F';
+    $cmdFormat = '%E -T%F %t -o ' . rtrim($argv[1], DIRECTORY_SEPARATOR) . '/configuration.graphviz.%F';
 } else {
     $cmdFormat = '';
 }