+++ examples/public-methods/graphviz.php
@@ -1,34 +1,36 @@
-<?php
-/**
- * Examples of class diagram on sources loaded with custom autoloader
- *
- * PHP version 7
- *
- * @category PHP
- * @package  UmlWriter
- * @author   Laurent Laville <pear@laurent-laville.org>
- * @license  https://opensource.org/licenses/BSD-3-Clause  The 3-Clause BSD License
- */
+<?php declare(strict_types=1);
 
-require_once 'autoloader/classmap.php';
+require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
 
+use Bartlett\GraphUml\Generator\GraphVizGenerator;
 use Bartlett\UmlWriter\Generator\GeneratorFactory;
 use Bartlett\UmlWriter\Service\ClassDiagramRenderer;
+
 use Symfony\Component\Finder\Finder;
 
 // path to directory where to find PHP source code
-$dataSource = __DIR__ . '/autoloader';
+$dataSource = dirname(__DIR__, 2) . '/src';
 
 $finder = new Finder();
 $finder->in($dataSource)->name('*.php');
 
 $generatorFactory = new GeneratorFactory('graphviz');
+
 // creates instance of Bartlett\GraphUml\Generator\GraphVizGenerator
+/** @var GraphVizGenerator $generator */
 $generator = $generatorFactory->getGenerator();
 
 $renderer = new ClassDiagramRenderer();
-// generates UML class diagram of all objects found in dataSource
-$script = $renderer($finder, $generator);
+$options = [
+    'show_private' => false,
+    'show_protected' => false,
+    'show_constants' => false,
+    'show_properties' => false,
+    'node.fillcolor' => '#FEFECE',
+    'node.style' => 'filled',
+];
+// generates UML class diagram of all objects found in dataSource (in graphviz format)
+$script = $renderer($finder, $generator, $options);
 // show UML diagram statements
 echo $script;
 
@@ -37,7 +39,7 @@ $generator->setFormat('svg');
 
 if (isset($argv[1])) {
     // target folder provided
-    $cmdFormat = '%E -T%F %t -o ' . rtrim($argv[1], DIRECTORY_SEPARATOR) . '/custom_autoloader.graphviz.%F';
+    $cmdFormat = '%E -T%F %t -o ' . rtrim($argv[1], DIRECTORY_SEPARATOR) . '/public_methods_only.graphviz.%F';
 } else {
     $cmdFormat = '';
 }